requirements-llm.txt
# LLM & RAG Project Dependencies
# Install: pip install -r requirements-llm.txt

# LangChain
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0

# LlamaIndex
llama-index>=0.11.0
llama-index-llms-openai>=0.2.0
llama-index-llms-anthropic>=0.2.0
llama-index-llms-huggingface>=0.2.0
llama-index-embeddings-openai>=0.2.0
llama-index-embeddings-huggingface>=0.2.0

# LLM API clients (optional, LangChain/LlamaIndex wrap these)
openai>=1.0.0
anthropic>=0.39.0
huggingface-hub>=0.26.0
transformers>=4.45.0
sentence-transformers>=3.0.0

# RAG & embeddings
chromadb>=0.5.0
tiktoken>=0.7.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
